requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
python-dotenv
openai>=1.0.0
httpx>=0.27.0
faiss-cpu

